np.int and np.bool are obsolete

+++ py/legacyhalos/hsc.py
@@ -188,7 +188,7 @@ def mpi_args():
 
     parser.add_argument('--first', type=int, help='Index of first object to process.')
     parser.add_argument('--last', type=int, help='Index of last object to process.')
-    parser.add_argument('--galaxylist', type=np.int, nargs='*', default=None, help='List of galaxy names to process.')
+    parser.add_argument('--galaxylist', type=int, nargs='*', default=None, help='List of galaxy names to process.')
 
     parser.add_argument('--coadds', action='store_true', help='Build the custom coadds.')
     parser.add_argument('--pipeline-coadds', action='store_true', help='Build the pipelinecoadds.')
@@ -454,7 +454,7 @@ def _build_multiband_mask(data, tractor, filt2pixscale, fill_value=0.0,
         mge, centralmask = tractor2mge(central, factor=neighborfactor)
         #plt.clf() ; plt.imshow(centralmask, origin='lower') ; plt.savefig('junk-mask.png') ; pdb.set_trace()
         
-        iclose = np.where([centralmask[np.int(by), np.int(bx)]
+        iclose = np.where([centralmask[int(by), int(bx)]
                            for by, bx in zip(tractor.by, tractor.bx)])[0]
         
         srcs = tractor.copy()
